+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Move report settings into a modal
 - Use HTML table for building grid and vgrid
 - Use PivotTable.js instead of cynteka for building pivot table
+- Hide dashboard title on print
 
 ### Removed